Changed enum state and logic; added tests.
After a lot of thinking, I decided to rename the states of walls
into s_passage, s_wall, and s_hard_wall. s_hard_wall is used for
the borders and for the walls that are made with the stamp method.
The removeCrosses method now can also deal with s_hard_wall,
although it may fail in some situations (I have not come up with
an example yet).
I added tests, which also give examples of ways inwhich the
methods can be called. Not all maze generating methods can be
used in combination with the stamp method.

Signed-off-by: FransFaase <faase009@planet.nl>